Custom name for Event Mode. Colors for songs
- Added the possibility to define a custom name for records stored when
playing in Event Mode, so it can be something different from EVNT. Just
change the value of **EventModeName**. Warning, it doesn't have a
defined length.

- Borrowed from 3.9+ the functionality to set a specific color for
songs (it'll be reflected wherever the song is shown: in the musicwheel,
in the records screen, etc). Like in 3.9+, it's the **#MENUCOLOR** tag
in the DWI/SM file.