+2 new things, reverted two
New stuff (all work by kriz):

- Now a custom image can be used for the LoadingWindow IN WINDOWS (Linux
is unchanged for now).
- Changes in how lights work. New preference and behavior:
-- Preference "BlinkGameButtonLightsOnPressInAttract" makes platforms
light up when the player presses them during attract mode.
If the preference doesn't exist or is set to false, the platforms will
turn on in inverse  sequence as the bass lights

Reverts/Changes:
- "Reverted" the assert I had commented before because it made Edit mode
crash. Instead of that I put in place a check if the actual game mode is
Oni.
- Reverted the "prevent combo announcer repeat" feature. ScreenHowToPlay
crashes with this.